refactor(composer): share one Composer across channels, DMs, and threads
Threads previously had a stripped-down composer (no attachment upload, no @mention menu) because it was a separate early copy of the channel composer that never kept up — two composers guaranteed drift. This collapses both into one shared web/src/views/Composer.tsx.

- New Composer: text + attachment upload (button / paste / drag-drop with progress) + @mention autocomplete + send, on the existing .composer styles.
- Chat.tsx channel/DM composer and ThreadPanel's inline composer both now render <Composer> (Chat.tsx: -128 lines). allowAsTask gates the As Task toggle (channels/DMs only; a thread reply is never a task). The agent wake-state hint is passed as topSlot.
- Wired up the previously-dead 'uploading' state so upload buttons disable mid-upload (the disabled={uploading} binding existed but was never set).
- Docs: ARCHITECTURE codemap (Chat.tsx + new Composer.tsx) + FEATURES entry.

Verified: typecheck (root+web) green; browser — channel @menu + send→POST 200 + render; thread now shows upload buttons + @menu, no As Task.
