+++ ARCHITECTURE.md
@@ -86,7 +86,8 @@ Lets any machine join a server without cloning the repo: `npx @open-tag/daemon -
 - `main.tsx` / `store.tsx` — Router entry / global state Context (dev-login → JWT, socket.io real-time subscription, REST wrappers; multi-server: bootstrap selects current workspace by `/s/:slug`, exposes `servers` / `capabilities` / `createServer`). Root `/` = public landing page (`views/Landing.tsx`); `*` catch-all `RootRedirect` → after dev-login, redirects to `/s/:slug`.
 - `ServerSwitcher.tsx` — Top-left brand = workspace switcher (lists `GET /servers`, switch, create).
 - `views/Auth.tsx` — Register/login page (`/login` · `/register`) + invite landing page (`/join/:token`); standalone from StoreProvider bootstrap, self-fetches `/api/auth/*`, on success stores `open-tag.token` and hard-navigates into the main app. Bootstrap prefers `open-tag.token`; falls back to dev-login (`?as=`).
-- `views/Chat.tsx` — Main group-chat view (multi-channel sidebar + DM + @completion menu + thread panel + reactions + attachments + as-task + files tab + embedded TaskBoard).
+- `views/Chat.tsx` — Main group-chat view (multi-channel sidebar + DM + thread panel + reactions + as-task + files tab + embedded TaskBoard). The message input is the shared `Composer` (below), used by channels, DMs, and threads alike.
+- `views/Composer.tsx` — Shared message composer for channels, DMs, and threads: text + attachment upload (button / paste / drag-drop, with progress) + @mention autocomplete + send. `allowAsTask` gates the As Task toggle (channels/DMs only; a thread reply is never a task).
 - `views/Members.tsx` — Agents grouped by machine + AgentProfile seven-tab panel (overview / activity / workspace / permissions / apps / DM / reminders) + create-agent modal.
 - `views/misc.tsx` — Tasks / Inbox / Computers / Search / Settings.
 - `views/Landing.tsx` + `landing/` — Public landing page (`/`): warm-editorial skin (tokens all scoped to `.lp-*`, strictly isolated from app editorial skin — no cross-contamination), hero / three-pillar / capability cards × 9 / engine / architecture / CTA / footer (9 sections). "Enter workspace" button navigates to `/s/:slug` when logged in, otherwise to `/login`. Copy and selling points are grounded in verified capabilities.
