+++ FEATURES.md
@@ -45,6 +45,7 @@
 - [x] **Inbox Mentions tab — message-grained @-activity stream**: the unified Inbox's Mentions filter is its own message-grained list (`GET /api/mentions`, backed by `mentions_target_idx`) of every message that @-mentions you, **read or unread** — each row deep-links straight to that message (`?msg=`) or opens the parent thread. Distinct from the channel-aggregated `all`/`unread` filters, which only badge an `@` while its channel still has unread (so an already-read `@` would otherwise drop off the Mentions view entirely). Verified: real send (`tester` @you → `message_mentions` row) → mark read → `/api/mentions` returns it `read:true` while the old `inbox?filter=mentions` reports `hasMention:false`; browser (Mentions tab lists the read @ + click jumps to the exact message). **Paginated** (`limit`/`offset` + `hasMore`, Saved-style "Load more" button — 50/page); verified by seeding 63 @s → first page 50 + Load more → all 63, then `hasMore:false`
 - [x] **Action Cards (B-mode quick-commit human-in-the-loop)**: agent lacks channel:create/agent:create scope → `open-tag action prepare --target <ch>` (stdin action JSON, variants channel:create/agent:create) sends proposal card → human opens **pre-filled creation modal** to review/edit/confirm → creates as themselves → `POST /api/actions/:id/mark-executed` marks "executed by X". Verified curl + browser end-to-end
 - [x] **Attachment enrichment**: human-side image inline + **fullscreen lightbox** (scroll-zoom/drag-pan/Esc), **video inline player** (`<video controls playsInline preload=metadata>` + codec fallback); agent-side `attachment view` **downloads bytes to local `attachments/`** (image → runtime vision). Browser verified
+- [x] **Unified message composer**: channels, DMs, and threads share one `Composer` component (`web/src/views/Composer.tsx`) — text input, attachment upload (button / paste / drag-drop with progress), and @mention autocomplete behave identically everywhere. Threads previously had a stripped-down composer (no uploads, no @ menu) and now match channels; **As Task stays channel/DM-only** (a thread reply is never a task). The channel composer's state + handlers moved out of `Chat.tsx` into `Composer`, removing the two-composer drift. Browser verified (channel: send + @menu; thread: upload buttons + @menu present; typecheck green)
 - [x] **Pluggable attachment storage**: `src/server/storage.ts` boundary — local disk (default, zero-config) or any **S3-compatible backend** (MinIO/Garage/SeaweedFS/OSS) via `OPEN_TAG_STORAGE=s3` + `OPEN_TAG_S3_*` (`@aws-sdk/client-s3` optional dep). Fail-loud config validation (`s3Config()` names the missing var). Bytes travel over HTTP so a remote daemon works cross-host. **Local-MinIO E2E verified**: human + agent round-trips byte-identical, missing-config → `500` + clear detail without crashing the server (regression test for the upload error-handling crash it surfaced)
 
 ## P5 Agent Collaboration — Core loop end-to-end verified (slice 01)
